Fix errors on non-Darwin platforms
HTTPURLResponse is part of FoundationNetworking instead of Foundation on non-Darwin platforms.
HTTPURLResponse.init is not accessible in FoundationNetworking, so init(url:statusCode:httpVersion:headerfields:) must be used to test for a missing date header.
HTTPURLResponse.value(forHTTPHeaderField:) is not present in the current release of FoundationNetworking, and allHeaderFields uses keys with different capitalization than the Foundation version. Fixes for both have been implemented in Swift 5.3.
Extraneous whitespace has been removed.